Moved converters into "internal" namespace
