+++ src/MaterialDesignThemes.Wpf/Converters/Internal/TextBoxHorizontalScrollBarWidthConverter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Globalization;
 using System.Windows.Data;
 
-namespace MaterialDesignThemes.Wpf.Converters;
+namespace MaterialDesignThemes.Wpf.Converters.Internal;
 
 public class TextBoxHorizontalScrollBarWidthConverter : IMultiValueConverter
 {
